+++ src/Konsole/Contracts/ControlStatus.cs
@@ -3,6 +3,7 @@
     public enum ControlStatus { 
         
         Undefined = 0, 
+
         /// <summary>
         /// control is active and has the focus and is in editing mode, pressing arrow keys moves the cursor.
         /// </summary>
@@ -10,7 +11,9 @@
 
         /// <summary>
         /// control has been selected but is still inactive, i.e. the form (and all controls) are in navigating mode. 
-        /// Pressing arrow keys will move the focus to another control.
+        /// Pressing arrow keys will move the focus to another control, pressing enter or starting to type
+        /// will send the keystrokes to the selected control. Before the keys are sent the control "Render" will 
+        /// be called updating the status.
         /// </summary>
         InactiveSelected = 2,