+++ src/Konsole/Drawing/Draw.cs
@@ -49,58 +49,6 @@ namespace Konsole
             return this;
         }
 
-
-        //public Draw Box(int sx, int sy, int ex, int ey, string title, LineThickNess? thicknessOverride = null)
-        //{
-        //    // TODO : remove duplication!
-
-        //    var thickness = thicknessOverride ?? _style.ThickNess;
-        //    int width = ex - sx + 1;
-        //    int height = ey - sy + 1;
-        //    // if box is not visible, return.
-        //    if (ex - sx < 0) return this;
-        //    if (ey - sy < 0) return this;
-
-        //    // if box is 1 character hight and wide render square and return
-        //    if (height == 1 && width == 1)
-        //    {
-        //        _console.PrintAt(sx, sy, '☐');
-        //        return this;
-        //    }
-        //    var line = thickness == LineThickNess.Single ? ThinBox : ThickBox;
-        //    DrawCorners(sx, sy, ex, ey, line);
-        //    // top edge
-        //    //Line(sx + 1, sy, ex - 1, sy, thickness);
-        //    DrawHorizontal(sx + 1, sy, ex - 1, line);
-
-        //    // left edge
-        //    //Line(sx, sy + 1, sx, ey - 1, thickness);
-        //    DrawVertical(sx, sy + 1, ey - 1, line);
-
-        //    // right edge
-        //    //Line(ex, sy + 1, ex, ey - 1, thickness);
-        //    DrawVertical(ex, sy + 1, ey - 1, line);
-
-        //    // bottom edge
-        //    //Line(sx + 1, ey, ex - 1, ey, thickness);
-        //    DrawHorizontal(sx + 1, ey, ex - 1, line);
-
-        //    // print centered title
-        //    var titleText = $" {title} ";
-        //    int len = titleText.Length;
-        //    int maxLen = width - 2;
-        //    if (len > maxLen)
-        //    {
-        //        titleText = maxLen > 0 ? titleText.Substring(0, maxLen) : "";
-        //    }
-        //    if (!string.IsNullOrWhiteSpace(titleText))
-        //    {
-        //        _console.PrintAt(sx + width / 2 - titleText.Length / 2, sy, titleText);
-        //    }
-
-        //    return this;
-        //}
-
         public Draw Box(int sx, int sy, int ex, int ey, string title)
         {
             return Box(sx, sy, ex, ey, title, null);