+++ src/Konsole/Menus/Menu.cs
@@ -249,7 +249,7 @@ namespace Konsole
         {
             throw new NotImplementedException();
         }
-        public override void Render(ControlStatus status, Style style)
+        protected override void Render(ControlStatus status, Style style)
         {
             lock (Window._locker)
             {